Calculating more fine grained decision timings

+++ hadoop-yarn-project/hadoop-yarn/hadoop-yarn-server/hadoop-yarn-server-resourcemanager/src/main/java/org/apache/hadoop/yarn/server/resourcemanager/scheduler/fair/FairScheduler.java
@@ -1027,7 +1027,7 @@ public class FairScheduler extends
     long duration = getClock().getTime() - start;
     fsOpDurations.addNodeUpdateDuration(duration);
     /* Start - Wajih*/
-    total_decision_time[duration]++;
+    total_decision_time[(int)duration]++;
     /* End */
   }
